requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
pathspec==0.7.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
